@@ -1,49 +1,51 @@
 {
-    "private": true,
-    "dependencies": {
-        "babel-core": "^6.14.0",
-        "babel-loader": "^6.2.5",
-        "babel-polyfill": "^6.13.0",
-        "babel-preset-es2015": "^6.14.0",
-        "babel-preset-react": "^6.11.1",
-        "html-webpack-plugin": "^2.22.0",
-        "lodash": "^4.15.0",
-        "react": "^15.3.1",
-        "react-dom": "^15.3.1",
-        "react-redux": "^4.4.5",
-        "react-router": "^2.8.1",
-        "redux": "^3.6.0",
-        "redux-thunk": "^2.1.0",
-        "webpack": "^1.13.2",
-        "webpack-split-by-path": "^0.1.0"
+  "private": true,
+  "dependencies": {
+    "babel-core": "^6.14.0",
+    "babel-loader": "^6.2.5",
+    "babel-polyfill": "^6.13.0",
+    "babel-preset-es2015": "^6.14.0",
+    "babel-preset-react": "^6.11.1",
+    "html-webpack-plugin": "^2.22.0",
+    "lodash": "^4.15.0",
+    "react": "^15.3.1",
+    "react-dom": "^15.3.1",
+    "react-redux": "^5.0.3",
+    "react-router-dom": "^4.0.0",
+    "redux": "^3.6.0",
+    "redux-thunk": "^2.1.0",
+    "webpack": "^2.3.3",
+    "webpack-split-by-path": "^2.0.0"
+  },
+  "devDependencies": {
+    "babel-jest": "^19.0.0",
+    "enzyme": "^2.4.1",
+    "eslint": "^3.6.0",
+    "eslint-loader": "^1.5.0",
+    "eslint-plugin-react": "^6.2.1",
+    "jest-cli": "^19.0.2",
+    "progress-bar-webpack-plugin": "^1.9.0",
+    "react-addons-test-utils": "^15.3.2"
+  },
+  "scripts": {
+    "build": "webpack --hide-modules",
+    "watch": "webpack --hide-modules --config webpack.dev.config.js",
+    "test": "jest",
+    "test:coverage": "jest --coverage"
+  },
+  "jest": {
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest"
     },
-    "devDependencies": {
-        "babel-jest": "^15.0.0",
-        "enzyme": "^2.4.1",
-        "eslint": "^3.6.0",
-        "eslint-loader": "^1.5.0",
-        "eslint-plugin-react": "^6.2.1",
-        "jest-cli": "^15.1.1",
-        "progress-bar-webpack-plugin": "^1.9.0",
-        "react-addons-test-utils": "^15.3.2"
-    },
-    "scripts": {
-        "build": "webpack",
-        "watch": "webpack --config webpack.dev.config.js",
-        "test": "jest",
-        "test:coverage": "jest --coverage"
-    },
-    "jest": {
-        "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
-        "testRegex": "/tests/.*\\.js$",
-        "testPathIgnorePatterns": [
-            "/node_modules/"
-        ],
-        "modulePaths": [
-            "src"
-        ],
-        "collectCoverageFrom": [
-            "src/*/**/*.js"
-        ]
-    }
+    "testMatch": ["**/tests/**/*.js"],
+    "testPathIgnorePatterns": [
+      "/node_modules/"
+    ],
+    "modulePaths": [
+      "src"
+    ],
+    "collectCoverageFrom": [
+      "src/*/**/*.js"
+    ]
+  }
 }
