@@ -5,15 +5,15 @@ import React from 'react';
 import Page from 'components/page.js';
 
 const Home = React.createClass({
-    render: function() {
-        return (
-            <Page>
-                <h1>
-                    Welcome
-                </h1>
-            </Page>
-        );
-    }
+  render: function() {
+    return (
+      <Page>
+        <h1>
+          Welcome
+        </h1>
+      </Page>
+    );
+  },
 });
 
 export default Home;
