@@ -1,12 +1,12 @@
 'use strict';
 
 import React from 'react';
-import {Route} from 'react-router';
+import {Route} from 'react-router-dom';
 
 import Home from 'components/home.js';
 import Async from 'containers/async.js';
 
 export default [
-    <Route path='/' component={Home}/>,
-    <Route path='/async' component={Async}/>
+  <Route path='/' component={Home}/>,
+  <Route path='/async' component={Async}/>,
 ];
