@@ -1,12 +1,12 @@
 'use strict';
 
 export function getWordOfTheDay() {
-    return new Promise((resolve) => {
-        setTimeout(() => {
-            resolve({
-                word: 'Hello!',
-                definition: 'Is it me you\'re looking for?'
-            });
-        }, 2000);
-    });
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve({
+        word: 'Hello!',
+        definition: 'Is it me you\'re looking for?',
+      });
+    }, 2000);
+  });
 }
