@@ -12,26 +12,34 @@ let config = require('./webpack.config.js');
 config.plugins = [];
 
 module.exports = _.merge({}, config, {
-    watch: true,
-    output: {
-        filename: '[name].js'
-    },
-    devtool: 'eval-source-map',
-    module: {
-        preLoaders: [{
-            test: /\.js$/,
-            exclude: /node_modules/,
-            loader: 'eslint-loader'
-        }]
-    },
-    plugins: [
-        new SplitByPathPlugin([{
-            name: 'vendor',
-            path: path.resolve('./node_modules')
-        }]),
-        new HtmlWebpackPlugin({
-            template: './assets/html/index.ejs'
-        }),
-        new ProgressBarPlugin()
-    ]
+  watch: true,
+  output: {
+    filename: '[name].js',
+  },
+  devtool: 'eval-source-map',
+  module: {
+    rules: [
+      {
+        enforce: 'pre',
+        test: /\.js$/,
+        exclude: /node_modules/,
+        loader: 'eslint-loader',
+      },
+      {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        loader: 'babel-loader',
+      },
+    ],
+  },
+  plugins: [
+    new SplitByPathPlugin([{
+      name: 'vendor',
+      path: path.resolve('./node_modules'),
+    }]),
+    new HtmlWebpackPlugin({
+      template: './assets/html/index.ejs',
+    }),
+    new ProgressBarPlugin(),
+  ],
 });
